requirements.txt
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langchain-chroma>=0.2.0
chromadb>=0.5.0
tavily-python>=0.5.0
yfinance>=0.2.40
gradio>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
